|#885|+Balvinder| Adding CHW model and controller API for creating a facility for CHW.  Changing type of parentFacilityId to Long.

+++ modules/core/src/main/java/org/openlmis/core/domain/Facility.java
@@ -110,7 +110,7 @@ public class Facility extends BaseModel implements Importable {
   private Boolean satellite;
 
   @ImportField(name = "Parent Facility ID")
-  private Integer parentFacilityId;
+  private Long parentFacilityId;
 
   @ImportField(name = "Facility Comments")
   private String comment;
